rename classes and table
