+++ server/odc-migrate/src/main/resources/migrate/common/V_4_3_0_2__add_object_search.sql
@@ -1,5 +1,5 @@
 -- Record all database objects such as table, view, function, procedure, package,  package_body, trigger, type, sequence, synonym and public synonym
-CREATE TABLE IF NOT EXISTS `connect_database_object` (
+CREATE TABLE IF NOT EXISTS `database_schema_object` (
   `id` bigint(20) NOT NULL AUTO_INCREMENT COMMENT 'ID of the database object',
   `name` varchar(128) NOT NULL COMMENT 'Name of the database object',
   `type` varchar(32) NOT NULL COMMENT 'Type of the database object, optional values: TABLE, VIEW, FUNCTION, PROCEDURE, PACKAGE, TRIGGER, TYPE, SEQUENCE, SYNONYM, PUBLIC_SYNONYM',
@@ -7,12 +7,12 @@ CREATE TABLE IF NOT EXISTS `connect_database_object` (
   `organization_id` bigint(20) NOT NULL COMMENT 'ID of the related organization, refer to iam_organization.id',
   `create_time` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Record insertion time',
   `update_time` datetime DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT 'Record modification time',
-  CONSTRAINT `pk_connect_database_object_id` PRIMARY KEY (`id`),
-  INDEX `idx_connect_database_object_database_id_type_name` (`database_id`, `type`, `name`)
+  CONSTRAINT `pk_database_schema_object_id` PRIMARY KEY (`id`),
+  INDEX `idx_database_schema_object_database_id_type_name` (`database_id`, `type`, `name`)
 ) COMMENT = 'Record all database objects such as table, view, function, procedure, package,  package_body, trigger, type, sequence, synonym and public synonym';
 
 -- Record all columns
-CREATE TABLE IF NOT EXISTS `connect_database_column` (
+CREATE TABLE IF NOT EXISTS `database_schema_column` (
   `id` bigint(20) NOT NULL AUTO_INCREMENT COMMENT 'ID of the column',
   `name` varchar(128) NOT NULL COMMENT 'Name of the column',
   `object_id` bigint(20) NOT NULL COMMENT 'ID of the related table or view, refer to connect_database_object.id',
@@ -20,8 +20,8 @@ CREATE TABLE IF NOT EXISTS `connect_database_column` (
   `organization_id` bigint(20) NOT NULL COMMENT 'ID of the related organization, refer to iam_organization.id',
   `create_time` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Record insertion time',
   `update_time` datetime DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT 'Record modification time',
-  CONSTRAINT `pk_connect_database_column_id` PRIMARY KEY (`id`),
-  CONSTRAINT `uk_connect_database_column_database_id_object_id_name` UNIQUE KEY (`database_id`, `object_id`, `name`)
+  CONSTRAINT `pk_database_schema_column_id` PRIMARY KEY (`id`),
+  CONSTRAINT `uk_database_schema_column_database_id_object_id_name` UNIQUE KEY (`database_id`, `object_id`, `name`)
 ) COMMENT = 'Record all columns';
 
 -- Add columns in `connect_database`
