+++ server/odc-service/src/main/java/com/oceanbase/odc/metadb/dbobject/DBColumnRepository.java
@@ -32,7 +32,7 @@ import com.oceanbase.odc.config.jpa.OdcJpaRepository;
 public interface DBColumnRepository extends OdcJpaRepository<DBColumnEntity, Long> {
 
     @Transactional
-    @Query(value = "select t.* from connect_database_column as t where t.database_id in (:databaseIds) and "
+    @Query(value = "select t.* from database_schema_column as t where t.database_id in (:databaseIds) and "
             + "t.name like :nameKey order by t.database_id desc, t.object_id desc LIMIT 1000;",
             nativeQuery = true)
     List<DBColumnEntity> findTop1000ByDatabaseIdInAndNameLike(@Param("databaseIds") Collection<Long> databaseIds,
