+++ server/odc-service/src/main/java/com/oceanbase/odc/service/db/schema/model/QueryDBObjectParams.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.db.object.model;
+package com.oceanbase.odc.service.db.schema.model;
 
 import java.util.List;
